requirements.txt
Django==2.0
django-ckeditor==5.7.1
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-widget-tweaks==1.4.5
Pillow==5.0.0
python-decouple==3.1
pytz==2019.1